fix: 修复评估价格float精度丢失问题
1. 将前端类型定义的eval_price改为支持number和string类型
2. 将后端相关字段从float替换为Decimal以避免精度损失
3. 简化Decimal转换逻辑，直接使用解析后的Decimal值

+++ backend/schemas/lead/__init__.py
@@ -1,6 +1,7 @@
 """Leads Management Pydantic Schemas."""
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -65,7 +66,7 @@ class PriceHistoryResponse(PriceHistoryBase):
 class LeadEvalHistoryCreate(BaseModel):
     """创建评估历史请求."""
 
-    eval_price: float = Field(gt=0, description="评估价格(万)")
+    eval_price: Decimal = Field(gt=0, decimal_places=2, description="评估价格(万)")
     remark: str | None = Field(None, max_length=500, description="评估备注")
